updated loan and portfolio testcase
